requirements.txt
Flask==2.2.2
psycopg2-binary
Werkzeug==2.3.7
flask-login